move analytics manager
