+++ lbrynet/analytics/__init__.py
@@ -1,6 +1,6 @@
+from constants import *
 from events import *
-from api import AnalyticsApi as Api
+from api import Api
 from track import Track
+from manager import Manager
 
-# Constants for metrics
-BLOB_BYTES_UPLOADED = 'blob_bytes_uploaded'
