+++ lbrynet/analytics/constants.py
@@ -1,0 +1,2 @@
+# Constants for metrics
+BLOB_BYTES_UPLOADED = 'blob_bytes_uploaded'
